Organize actions and api models
