+++ lib/api/shared/src/main/scala/net/wiringbits/api/models/users/ResetPassword.scala
@@ -1,5 +1,6 @@
-package net.wiringbits.api.models
+package net.wiringbits.api.models.users
 
+import net.wiringbits.api.models.*
 import net.wiringbits.common.models.{Email, Name, Password, UserToken}
 import play.api.libs.json.{Format, Json}
 import sttp.tapir.Schema
