+++ server/src/main/scala/net/wiringbits/actions/users/ResetPasswordAction.scala
@@ -1,6 +1,6 @@
-package net.wiringbits.actions
+package net.wiringbits.actions.users
 
-import net.wiringbits.api.models.ResetPassword
+import net.wiringbits.api.models.users.ResetPassword
 import net.wiringbits.common.models.Password
 import net.wiringbits.config.UserTokensConfig
 import net.wiringbits.repositories.{UserTokensRepository, UsersRepository}
