+++ server/src/main/scala/net/wiringbits/actions/users/SendEmailVerificationTokenAction.scala
@@ -1,6 +1,6 @@
-package net.wiringbits.actions
+package net.wiringbits.actions.users
 
-import net.wiringbits.api.models.SendEmailVerificationToken
+import net.wiringbits.api.models.users.SendEmailVerificationToken
 import net.wiringbits.apis.ReCaptchaApi
 import net.wiringbits.repositories.UsersRepository
 import net.wiringbits.util.EmailsHelper
