+++ server/src/test/scala/controllers/AdminControllerSpec.scala
@@ -2,7 +2,7 @@ package controllers
 
 import com.dimafeng.testcontainers.PostgreSQLContainer
 import controllers.common.PlayPostgresSpec
-import net.wiringbits.api.models.CreateUser
+import net.wiringbits.api.models.users.CreateUser
 import net.wiringbits.apis.models.EmailRequest
 import net.wiringbits.apis.{EmailApi, ReCaptchaApi}
 import net.wiringbits.common.models.*
