+++ server/src/test/scala/controllers/UsersControllerSpec.scala
@@ -2,14 +2,8 @@ package controllers
 
 import com.dimafeng.testcontainers.PostgreSQLContainer
 import controllers.common.PlayPostgresSpec
-import net.wiringbits.api.models.{
-  CreateUser,
-  ForgotPassword,
-  Login,
-  ResetPassword,
-  SendEmailVerificationToken,
-  VerifyEmail
-}
+import net.wiringbits.api.models.auth.Login
+import net.wiringbits.api.models.users.{CreateUser, ForgotPassword, ResetPassword, SendEmailVerificationToken, VerifyEmail}
 import net.wiringbits.apis.models.EmailRequest
 import net.wiringbits.apis.{EmailApi, ReCaptchaApi}
 import net.wiringbits.common.models.*
